+++ OUDS/Core/Components/Sources/Checkbox/Internal/CheckboxIndicator.swift
@@ -17,12 +17,12 @@ import OUDSTokensSemantic
 import SwiftUI
 
 /// The indicator of the checkbox.
-/// Its content depends to the ``ControlItemInternalState`` and the ``OUDSCheckboxIndicatorState`` also.
+/// Its content depends to the ``InteractionState`` and the ``OUDSCheckboxIndicatorState`` also.
 struct CheckboxIndicator: View {
 
     // MARK: - Properties
 
-    let internalState: ControlItemInternalState
+    let interactionState: InteractionState
     let indicatorState: OUDSCheckboxIndicatorState
     let isError: Bool
 
@@ -33,7 +33,9 @@ struct CheckboxIndicator: View {
 
     var body: some View {
         indicator()
-            .modifier(CheckboxIndicatorStyle(state: internalState, indicatorState: indicatorState, isError: isError))
+            .modifier(CheckboxIndicatorModifier(interactionState: interactionState,
+                                                indicatorState: indicatorState,
+                                                isError: isError))
     }
 
     // MARK: - Indicator
@@ -46,7 +48,6 @@ struct CheckboxIndicator: View {
             tickImage(name: "ic_form_dash")
         } else { // .unselected
             Color.clear
-                .modifier(IndicatorFrameModifier())
         }
     }
 
@@ -54,14 +55,13 @@ struct CheckboxIndicator: View {
         Image(decorative: name, bundle: Bundle.OUDSComponents)
             .resizable()
             .scaledToFit()
-            .modifier(IndicatorFrameModifier())
             .accessibilityHidden(true)
-            .foregroundColor(appliedColor.color(for: colorScheme))
+            .oudsForegroundColor(appliedColor)
     }
 
     private var appliedColor: MultipleColorSemanticTokens {
         if isError {
-            switch internalState {
+            switch interactionState {
             case .enabled:
                 return theme.colors.colorActionNegativeEnabled
             case .hover:
@@ -73,7 +73,7 @@ struct CheckboxIndicator: View {
                          + " Only non-error situation are allowed to have a disabled state / read only mode.")
             }
         } else {
-            switch internalState {
+            switch interactionState {
             case .enabled:
                 return theme.colors.colorActionSelected
             case .hover:
@@ -84,18 +84,5 @@ struct CheckboxIndicator: View {
                 return theme.colors.colorActionDisabled
             }
         }
-    }
-}
-
-// MARK: - Indicator Frame Modifier
-
-private struct IndicatorFrameModifier: ViewModifier {
-
-    @Environment(\.theme) private var theme
-
-    func body(content: Content) -> some View {
-        content
-            .frame(width: theme.checkbox.checkboxSizeIndicator,
-                   height: theme.checkbox.checkboxSizeIndicator)
     }
 }