+++ OUDS/Core/Components/Sources/Link/Internal/LinkSizeIconModifier.swift
@@ -18,8 +18,8 @@ struct LinkSizeIconModifier: ViewModifier {
     // MARK: - Properties
 
     let size: OUDSLink.Size
+
     @Environment(\.theme) private var theme
-    @Environment(\.colorScheme) private var colorScheme
 
     // MARK: - Body