+++ OUDS/Core/Components/Sources/Link/Internal/LinkTextModifier.swift
@@ -19,12 +19,11 @@ struct LinkTextModifier: ViewModifier {
 
     // MARK: - Properties
 
+    let interactionState: InteractionState
     let size: OUDSLink.Size
     let layout: OUDSLink.Layout
-    let state: LinkInternalState
 
     @Environment(\.theme) private var theme
-    @Environment(\.colorScheme) private var colorScheme
 
     // MARK: - Body
 
@@ -38,7 +37,7 @@ struct LinkTextModifier: ViewModifier {
             }
         }
         .multilineTextAlignment(.leading)
-        .modifier(LinkUnderlineModifier(layout: layout, state: state))
+        .modifier(LinkUnderlineModifier(interactionState: interactionState, layout: layout))
     }
 }
 
@@ -46,9 +45,10 @@ struct LinkTextModifier: ViewModifier {
 
 private struct LinkUnderlineModifier: ViewModifier {
 
-    @State private var textWidth: CGFloat = 0
+    let interactionState: InteractionState
     let layout: OUDSLink.Layout
-    let state: LinkInternalState
+
+    @State private var textWidth: CGFloat = 0
 
     func body(content: Content) -> some View {
         if #available(iOS 16.0, *) {
@@ -58,9 +58,7 @@ private struct LinkUnderlineModifier: ViewModifier {
                 VStack(spacing: 0) {
                     content
                         .readSize { size in
-                            Task { @MainActor in
-                                textWidth = size.width
-                            }
+                            textWidth = size.width
                         }
                     Rectangle().frame(width: textWidth, height: 1)
                 }
@@ -73,36 +71,11 @@ private struct LinkUnderlineModifier: ViewModifier {
     private var underlineActive: Bool {
         switch layout {
         case .arrow:
-            return state == .hover || state == .pressed
+            return interactionState == .hover || interactionState == .pressed
         case .textOnly:
             return true
         case .iconAndText:
-            return state == .hover || state == .pressed
+            return interactionState == .hover || interactionState == .pressed
         }
     }
 }
-
-// MARK: - Sizes management
-
-private struct SizePreferenceKey: @preconcurrency PreferenceKey {
-    @MainActor static var defaultValue: CGSize = .zero
-
-    static func reduce(value _: inout CGSize, nextValue _: () -> CGSize) {}
-}
-
-// swiftlint:disable strict_fileprivate
-extension View {
-
-    /// Use to read the size of the current view (usefull to read the width of text)
-    /// - Parameter onChange: Called when the size changes
-    fileprivate func readSize(onChange: @Sendable @escaping (CGSize) -> Void) -> some View {
-        background(
-            GeometryReader { geometryProxy in
-                Color.clear
-                    .preference(key: SizePreferenceKey.self, value: geometryProxy.size)
-            }
-        )
-        .onPreferenceChange(SizePreferenceKey.self, perform: onChange)
-    }
-}
-// swiftlint:enable strict_fileprivate