+++ OUDS/Core/Components/Sources/Radio/Internal/RadioIndicatorModifiers.swift
@@ -16,15 +16,15 @@ import OUDSFoundations
 import OUDSTokensSemantic
 import SwiftUI
 
-// MARK: - Radio Indicator Style
+// MARK: - Radio Indicator Modifier
 
 /// A `ViewModier` to apply to the ``RadioIndicator`` component.
-/// It will define the look and feel of the indicator depending to the ``ControlItemInternalState`` and some flags
-struct RadioIndicatorStyle: ViewModifier {
+/// It will define the look and feel of the indicator depending to the ``InteractionState`` and some flags
+struct RadioIndicatorModifier: ViewModifier {
 
     // MARK: - Properties
 
-    let state: ControlItemInternalState
+    let interactionState: InteractionState
     let isOn: Bool
     let isError: Bool
 
@@ -32,9 +32,10 @@ struct RadioIndicatorStyle: ViewModifier {
 
     func body(content: Content) -> some View {
         content
-            .modifier(RadioIndicatorBorderModifier(state: state, isOn: isOn, isError: isError))
-            .modifier(RadioIndicatorForegroundModifier(state: state, isOn: isOn, isError: isError))
-            .modifier(RadioIndicatorBackgroundModifier(state: state, isOn: isOn, isError: isError))
+            .modifier(SizeFrameModifier())
+            .modifier(RadioIndicatorBorderModifier(interactionState: interactionState, isOn: isOn, isError: isError))
+            .modifier(RadioIndicatorForegroundModifier(interactionState: interactionState, isOn: isOn, isError: isError))
+            .modifier(RadioIndicatorBackgroundModifier(interactionState: interactionState, isOn: isOn, isError: isError))
     }
 }
 
@@ -44,7 +45,7 @@ private struct RadioIndicatorForegroundModifier: ViewModifier {
 
     // MARK: - Properties
 
-    let state: ControlItemInternalState
+    let interactionState: InteractionState
     let isOn: Bool
     let isError: Bool
 
@@ -55,21 +56,21 @@ private struct RadioIndicatorForegroundModifier: ViewModifier {
 
     func body(content: Content) -> some View {
         content
-            .foregroundColor(appliedColor)
+            .oudsForegroundColor(appliedColor)
     }
 
     // MARK: - Colors
 
-    private var appliedColor: Color {
-        switch state {
+    private var appliedColor: MultipleColorSemanticTokens {
+        switch interactionState {
         case .enabled:
-            return enabledColor.color(for: colorScheme)
+            return enabledColor
         case .hover:
-            return hoverColor.color(for: colorScheme)
+            return hoverColor
         case .pressed:
-            return pressedColor.color(for: colorScheme)
+            return pressedColor
         case .disabled, .readOnly:
-            return disabledColor.color(for: colorScheme)
+            return disabledColor
         }
     }
 
@@ -104,7 +105,7 @@ private struct RadioIndicatorBackgroundModifier: ViewModifier {
 
     // MARK: - Properties
 
-    let state: ControlItemInternalState
+    let interactionState: InteractionState
     let isOn: Bool
     let isError: Bool
 
@@ -121,7 +122,7 @@ private struct RadioIndicatorBackgroundModifier: ViewModifier {
     // MARK: - Colors
 
     private var appliedColor: Color {
-        switch state {
+        switch interactionState {
         case .enabled:
             return enabledColor
         case .hover:
@@ -160,7 +161,7 @@ private struct RadioIndicatorBorderModifier: ViewModifier {
 
     // MARK: - Properties
 
-    let state: ControlItemInternalState
+    let interactionState: InteractionState
     let isOn: Bool
     let isError: Bool
 
@@ -180,7 +181,7 @@ private struct RadioIndicatorBorderModifier: ViewModifier {
     // MARK: - Colors
 
     private var appliedColor: MultipleColorSemanticTokens {
-        switch state {
+        switch interactionState {
         case .enabled:
             return enabledColor
         case .hover:
@@ -227,7 +228,7 @@ private struct RadioIndicatorBorderModifier: ViewModifier {
     // MARK: - Border width
 
     private var appliedBorderWidth: CGFloat {
-        switch state {
+        switch interactionState {
         case .enabled:
             return enabledWidth
         case .hover:
@@ -259,5 +260,18 @@ private struct RadioIndicatorBorderModifier: ViewModifier {
 
     private var appliedBorderRadius: CGFloat {
         theme.radioButton.radioButtonBorderRadius
+    }
+}
+
+// MARK: - Indicator Frame Modifier
+
+private struct SizeFrameModifier: ViewModifier {
+
+    @Environment(\.theme) private var theme
+
+    func body(content: Content) -> some View {
+        content
+            .frame(width: theme.radioButton.radioButtonSizeIndicator,
+                   height: theme.radioButton.radioButtonSizeIndicator)
     }
 }